requirements.txt
asgiref==3.5.2
dj-database-url==0.5.0
Django==4.0.6
django-silk==5.0.1
django-solo==2.0.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
django-cors-headers==3.13.0
django-extensions==3.2.0
drf-yasg==1.20.0
gunicorn==20.1.0
psycopg2==2.9.3
pytz==2022.1
pyyaml==6.0.0
six==1.16.0
sqlparse==0.4.2
tblib==1.7.0
whitenoise==6.2.0
